+++ contracts/CryptoravesAdministration/Ravepool.sol
@@ -6,26 +6,26 @@ import "/home/cartosys/openzeppelin-contracts/contracts/token/ERC1155/ERC1155Bur
 import "./AdministrationContract.sol";
 
 interface CryptoravesTokenContract {
-    function cryptoravesIdByAddress(address _account) external view returns(uint256);
-    function getHeldTokenIds(address _addr) external view returns(uint256[] memory);
-    function getHeldTokenBalances(address _addr) external view returns(uint256[] memory);
-    function safeTransferFrom(address from, address to, uint256 id, uint256 amount, bytes memory data) external;
-    function safeBatchTransferFrom(address from, address to, uint256[] memory ids, uint256[] memory amounts, bytes memory data) external;
-    function burn(address, uint256, uint256) external;
-    function subtractFromTotalSupply(uint256 _tokenId, uint256 _amount) external;
+    function cryptoravesIdByAddress(address _account) external view returns(uint);
+    function getHeldTokenIds(address _addr) external view returns(uint[] memory);
+    function getHeldTokenBalances(address _addr) external view returns(uint[] memory);
+    function safeTransferFrom(address from, address to, uint id, uint amount, bytes memory data) external;
+    function safeBatchTransferFrom(address from, address to, uint[] memory ids, uint[] memory amounts, bytes memory data) external;
+    function burn(address, uint, uint) external;
+    function subtractFromTotalSupply(uint _tokenId, uint _amount) external;
 }
 
 
 contract Ravepool is AdministrationContract {
 
-    using SafeMath for uint256;
+    using SafeMath for uint;
     using Address for address;
 
     bool private _ravepoolActivated;
     address internal _tokenManager;
-    uint256 private _userId;
+    uint private _userId;
 
-    event BurnAndRedeem(uint256 _sentTokenId, uint256 _amountOfPersonalToken);
+    event BurnAndRedeem(uint _sentTokenId, uint _amountOfPersonalToken);
 
     modifier ravepoolActivated () {
       require(_ravepoolActivated, 'Ravepool is not activated. Reverting.');
@@ -42,7 +42,7 @@ contract Ravepool is AdministrationContract {
         address _userTokenManager = ITransactionManager(getTransactionManagerAddress()).userManagementContractAddress();
 
         //require crypto already dropped by user
-        uint256 _userIdToCheck = IUserManager(_userTokenManager).getUserId(address(this));
+        uint _userIdToCheck = IUserManager(_userTokenManager).getUserId(address(this));
         require(
             IUserManager(_userTokenManager).dropState(_userIdToCheck),
             'User has not yet dropped their cryptoraves personal token.'
@@ -61,27 +61,27 @@ contract Ravepool is AdministrationContract {
         return _ravepoolActivated;
     }
 
-    function redeemAndBurnViaRavepool(uint256 _sentTokenId, uint256 _amountOfPersonalToken, bytes memory _data) public payable onlyAdmin ravepoolActivated {
+    function redeemAndBurnViaRavepool(uint _sentTokenId, uint _amountOfPersonalToken, bytes memory _data) public payable onlyAdmin ravepoolActivated {
 
         require(_amountOfPersonalToken > 0, 'Amount must be greater than zero');
 
         address _cryptoravesTokenAddress = ITokenManager(_tokenManager).cryptoravesTokenAddress();
 
         //get personal token Id
-        uint256 _1155tokenId  = CryptoravesTokenContract(_cryptoravesTokenAddress).cryptoravesIdByAddress(address(this));
+        uint _1155tokenId  = CryptoravesTokenContract(_cryptoravesTokenAddress).cryptoravesIdByAddress(address(this));
 
         //require sent token to be the designated personal token
         require(_sentTokenId == _1155tokenId, 'Token ID sent to redeemAndBurnViaRavepool doesn\'t match designated burn token ID');
 
         //gather list of all held tokens in Ravepool
-        uint256[] memory _heldTokenIds = CryptoravesTokenContract(_cryptoravesTokenAddress).getHeldTokenIds(address(this));
-        uint256[] memory _heldTokenbalances = CryptoravesTokenContract(_cryptoravesTokenAddress).getHeldTokenBalances(address(this));
+        uint[] memory _heldTokenIds = CryptoravesTokenContract(_cryptoravesTokenAddress).getHeldTokenIds(address(this));
+        uint[] memory _heldTokenbalances = CryptoravesTokenContract(_cryptoravesTokenAddress).getHeldTokenBalances(address(this));
 
         //total supply used to calculate final share percentaghe
-        uint256 _totalSupplyOfPersonalToken = ITokenManager(_tokenManager).managedTokenByFullBytesId(_sentTokenId).totalSupply;
+        uint _totalSupplyOfPersonalToken = ITokenManager(_tokenManager).managedTokenByFullBytesId(_sentTokenId).totalSupply;
 
         //calculate percentage of each token to be distributed back to msg sender
-        uint256[] memory _distributionAmounts;
+        uint[] memory _distributionAmounts;
 
         //will have to ensure erctype is 20 (fungible) for each of these. Then create another function to facilitate non-fungibles
         for (uint i=0; i<_heldTokenIds.length; i++) {
