+++ contracts/CryptoravesAdministration/WalletFull.sol
@@ -19,7 +19,7 @@ interface IERC1155Mintable {
 
 contract WalletFull is ERC1155Receiver {
 
-    using SafeMath for uint256;
+    using SafeMath for uint;
     using Address for address;
 
     address private _transactionManagerAddress;
@@ -35,30 +35,30 @@ contract WalletFull is ERC1155Receiver {
         setAdministrator(_txnManagerAddress);
     }
 
-    function onERC1155Received(address, address, uint256, uint256, bytes calldata) external override virtual returns (bytes4) {
+    function onERC1155Received(address, address, uint, uint, bytes calldata) external override virtual returns (bytes4) {
         return this.onERC1155Received.selector;
     }
 
-    function onERC1155BatchReceived(address, address, uint256[] calldata, uint256[] calldata, bytes calldata) external override virtual returns (bytes4) {
+    function onERC1155BatchReceived(address, address, uint[] calldata, uint[] calldata, bytes calldata) external override virtual returns (bytes4) {
         return this.onERC1155BatchReceived.selector;
     }
 /*
-    function managedTransfer(address _from, address _to, uint256 _id,  uint256 _val, bytes memory _data) public onlyAdmin {
+    function managedTransfer(address _from, address _to, uint _id,  uint _val, bytes memory _data) public onlyAdmin {
         address _cryptoravesTokenAddress = ITokenManager(_transactionManagerAddress).cryptoravesTokenAddress();
         IERC1155(_cryptoravesTokenAddress).safeTransferFrom(_from, _to, _id, _val, _data);
     }
 
-    function managedBatchTransfer(address _from, address _to, uint256[] memory _ids,  uint256[] memory _vals, bytes memory _data) public onlyAdmin {
+    function managedBatchTransfer(address _from, address _to, uint[] memory _ids,  uint[] memory _vals, bytes memory _data) public onlyAdmin {
         address _cryptoravesTokenAddress = ITokenManager(_transactionManagerAddress).cryptoravesTokenAddress();
         IERC1155(_cryptoravesTokenAddress).safeBatchTransferFrom(_from, _to, _ids, _vals, _data);
     }
 
-    function managedBurn(address account, uint256 id, uint256 amount) public onlyAdmin {
+    function managedBurn(address account, uint id, uint amount) public onlyAdmin {
         address _cryptoravesTokenAddress = ITokenManager(_transactionManagerAddress).cryptoravesTokenAddress();
         ERC1155Burnable(_cryptoravesTokenAddress).burn(account, id, amount);
     }
 
-    function managedBurnBatch(address account, uint256[] memory ids, uint256[] memory amounts) public onlyAdmin {
+    function managedBurnBatch(address account, uint[] memory ids, uint[] memory amounts) public onlyAdmin {
         address _cryptoravesTokenAddress = ITokenManager(_transactionManagerAddress).cryptoravesTokenAddress();
         ERC1155Burnable(_cryptoravesTokenAddress).burnBatch(account, ids, amounts);
     }
